Thanks Michael Whitlock (Bear)!
ColorUtils.channelInt() was causing trouble in Bear's Java-parsing code, and with good reason -- it was implicitly casting its correct int result to a float, despite being named already as returning an int. So, now channel() returns a float, channelInt() returns an int.


Former-commit-id: 576672e93795d372b12f2eec4d7052d2c1258a56

+++ src/main/java/com/github/tommyettinger/textra/utils/ColorUtils.java
@@ -158,7 +158,7 @@ public class ColorUtils {
      * @param channel which channel to access, as an index from 0 to 3 inclusive
      * @return the int value of the requested channel, from 0 to 255 inclusive
      */
-    public static float channelInt(int color, int channel) {
+    public static int channelInt(int color, int channel) {
         return (color >>> 24 - ((channel & 3) << 3) & 255);
     }
